Moved SchemaCacheInvalidationTest into right directory.
